+++ examples/demo_async_tokio.rs
@@ -1,7 +1,5 @@
-use delay_timer::prelude::*;
-
 use anyhow::Result;
-
+use delay_timer::prelude::*;
 use smol::Timer;
 use std::time::Duration;
 
@@ -54,7 +52,7 @@ fn build_task_async_print() -> Result<Task, TaskError> {
 
     task_builder
         .set_task_id(1)
-        .set_frequency(Frequency::Repeated("*/6 * * * * * *"))
+        .set_frequency_repeated_by_seconds(6)
         .set_maximum_parallel_runnable_num(2)
         .spawn(body)
 }
@@ -64,7 +62,7 @@ fn build_task_async_execute_process() -> Result<Task, TaskError> {
 
     let body = unblock_process_task_fn("php /home/open/project/rust/repo/myself/delay_timer/examples/try_spawn.php >> ./try_spawn.txt".into());
     task_builder
-        .set_frequency_by_candy(CandyFrequency::Repeated(CandyCron::Secondly))
+        .set_frequency_repeated_by_seconds(1)
         .set_task_id(3)
         .set_maximum_running_time(10)
         .set_maximum_parallel_runnable_num(1)