+++ src/timer/task.rs
@@ -54,21 +54,28 @@ impl TaskMark {
 
     #[inline(always)]
     pub(crate) fn set_parallel_runnable_num(&mut self, parallel_runnable_num: u64) -> &mut Self {
+        debug!(
+            "task-id: {}, parallel_runnable_num: {}",
+            self.task_id, self.parallel_runnable_num
+        );
         self.parallel_runnable_num = parallel_runnable_num;
         self
     }
 
     #[inline(always)]
     pub(crate) fn inc_parallel_runnable_num(&mut self) {
-        self.parallel_runnable_num += 1;
+        let parallel_runnable_num = self.parallel_runnable_num + 1;
+        self.set_parallel_runnable_num(parallel_runnable_num);
     }
 
     #[inline(always)]
     pub(crate) fn dec_parallel_runnable_num(&mut self) {
-        self.parallel_runnable_num = self
+        let parallel_runnable_num = self
             .parallel_runnable_num
             .checked_sub(1)
             .unwrap_or_default();
+
+        self.set_parallel_runnable_num(parallel_runnable_num);
     }
 
     #[inline(always)]
@@ -131,30 +138,122 @@ impl TaskMark {
 }
 
 #[derive(Debug, Copy, Clone)]
-/// Enumerated values of repeating types.
-pub enum Frequency<'a> {
+pub(crate) enum FrequencyUnify<'a> {
+    FrequencyCronStr(FrequencyCronStr<'a>),
+    FrequencySeconds(FrequencySeconds),
+}
+
+impl<'a> Default for FrequencyUnify<'a> {
+    fn default() -> FrequencyUnify<'a> {
+        FrequencyUnify::FrequencySeconds(FrequencySeconds::default())
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
+/// Enumerated values of repeating types based on the string of cron-expression.
+pub enum FrequencyCronStr<'a> {
     /// Repeat once.
     Once(&'a str),
     /// Repeat ad infinitum.
     Repeated(&'a str),
     /// Type of countdown.
-    CountDown(u32, &'a str),
+    CountDown(u64, &'a str),
 }
 
-impl<'a> Default for Frequency<'a> {
-    fn default() -> Frequency<'a> {
-        Frequency::Once("@minutely")
+#[derive(Debug, Copy, Clone)]
+/// Enumerated values of repeating types based on the number of seconds.
+pub(crate) enum FrequencySeconds {
+    /// Repeat once.
+    Once(u64),
+    /// Repeat ad infinitum.
+    Repeated(u64),
+    /// Type of countdown.
+    CountDown(u64, u64),
+}
+
+impl<'a> Default for FrequencyCronStr<'a> {
+    fn default() -> FrequencyCronStr<'a> {
+        FrequencyCronStr::Once("@minutely")
+    }
+}
+
+impl Default for FrequencySeconds {
+    fn default() -> FrequencySeconds {
+        FrequencySeconds::Once(ONE_MINUTE)
     }
 }
 
 /// Iterator for task internal control of execution time.
 #[derive(Debug, Clone)]
 pub(crate) enum FrequencyInner {
-    ///Unlimited repetition types.
-    Repeated(DelayTimerScheduleIteratorOwned),
-    ///Type of countdown.
-    CountDown(u32, DelayTimerScheduleIteratorOwned),
+    /// Unlimited repetition types for cron-expression.
+    CronExpressionRepeated(DelayTimerScheduleIteratorOwned),
+    /// Type of countdown for cron-expression.
+    CronExpressionCountDown(u64, DelayTimerScheduleIteratorOwned),
+    /// Unlimited repetition types for seconds-duration.
+    SecondsRepeated(SecondsState),
+    /// Type of countdown for SecondsState.
+    /// SecondsCountDown(count_down, SecondsState)
+    SecondsCountDown(u64, SecondsState),
 }
+
+impl<'a> TryFrom<(FrequencyUnify<'a>, ScheduleIteratorTimeZone)> for FrequencyInner {
+    type Error = FrequencyAnalyzeError;
+
+    fn try_from(
+        (frequency, time_zone): (FrequencyUnify<'_>, ScheduleIteratorTimeZone),
+    ) -> Result<FrequencyInner, Self::Error> {
+        let frequency_inner = match frequency {
+            FrequencyUnify::FrequencyCronStr(FrequencyCronStr::Once(cron_str)) => {
+                let task_schedule =
+                    DelayTimerScheduleIteratorOwned::analyze_cron_expression(time_zone, cron_str)?;
+
+                FrequencyInner::CronExpressionCountDown(1, task_schedule)
+            }
+            FrequencyUnify::FrequencyCronStr(FrequencyCronStr::Repeated(cron_str)) => {
+                let task_schedule =
+                    DelayTimerScheduleIteratorOwned::analyze_cron_expression(time_zone, cron_str)?;
+
+                FrequencyInner::CronExpressionRepeated(task_schedule)
+            }
+            FrequencyUnify::FrequencyCronStr(FrequencyCronStr::CountDown(count_down, cron_str)) => {
+                let task_schedule =
+                    DelayTimerScheduleIteratorOwned::analyze_cron_expression(time_zone, cron_str)?;
+
+                FrequencyInner::CronExpressionCountDown(count_down as u64, task_schedule)
+            }
+
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::Once(seconds)) => {
+                if seconds == 0 {
+                    return Err(FrequencyAnalyzeError::DisInitTime);
+                }
+
+                let seconds_state: SecondsState = (get_timestamp()..).step_by(seconds as usize);
+                FrequencyInner::SecondsCountDown(1, seconds_state)
+            }
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::Repeated(seconds)) => {
+                if seconds == 0 {
+                    return Err(FrequencyAnalyzeError::DisInitTime);
+                }
+
+                let seconds_state: SecondsState = (get_timestamp()..).step_by(seconds as usize);
+
+                FrequencyInner::SecondsRepeated(seconds_state)
+            }
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::CountDown(count_down, seconds)) => {
+                if seconds == 0 {
+                    return Err(FrequencyAnalyzeError::DisInitTime);
+                }
+
+                let seconds_state: SecondsState = (get_timestamp()..).step_by(seconds as usize);
+                FrequencyInner::SecondsCountDown(count_down, seconds_state)
+            }
+        };
+
+        Ok(frequency_inner)
+    }
+}
+
 /// Set the time zone for the time of the expression iteration.
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
 pub enum ScheduleIteratorTimeZone {
@@ -259,7 +358,7 @@ impl DelayTimerScheduleIteratorOwned {
     fn analyze_cron_expression(
         time_zone: ScheduleIteratorTimeZone,
         cron_expression: &str,
-    ) -> Result<DelayTimerScheduleIteratorOwned, CronExpressionAnalyzeError> {
+    ) -> Result<DelayTimerScheduleIteratorOwned, FrequencyAnalyzeError> {
         let indiscriminate_expression = cron_expression.trim_matches(' ').to_owned();
         let schedule_iterator_time_zone_query: ScheduleIteratorTimeZoneQuery =
             ScheduleIteratorTimeZoneQuery {
@@ -292,34 +391,45 @@ impl DelayTimerScheduleIteratorOwned {
 }
 
 impl FrequencyInner {
-    //How many times the acquisition needs to be performed.
+    // How many times the acquisition needs to be performed.
     #[allow(dead_code)]
-    fn residual_time(&self) -> u32 {
+    fn residual_time(&self) -> u64 {
         match self {
-            FrequencyInner::Repeated(_) => u32::MAX,
-            FrequencyInner::CountDown(ref time, _) => *time,
+            FrequencyInner::CronExpressionRepeated(_) => u64::MAX,
+            FrequencyInner::SecondsRepeated(_) => u64::MAX,
+            FrequencyInner::CronExpressionCountDown(ref time, _) => *time,
+            FrequencyInner::SecondsCountDown(ref time, _) => *time,
         }
     }
 
     fn next_alarm_timestamp(&mut self) -> Option<i64> {
         match self {
-            FrequencyInner::CountDown(_, ref mut clock) => clock.next(),
-            FrequencyInner::Repeated(ref mut clock) => clock.next(),
+            FrequencyInner::CronExpressionCountDown(_, ref mut clock) => clock.next(),
+            FrequencyInner::CronExpressionRepeated(ref mut clock) => clock.next(),
+            FrequencyInner::SecondsRepeated(seconds_state) => {
+                seconds_state.next().map(|s| s as i64)
+            }
+            FrequencyInner::SecondsCountDown(_, seconds_state) => {
+                seconds_state.next().map(|s| s as i64)
+            }
         }
     }
 
     #[warn(unused_parens)]
     fn down_count(&mut self) {
         match self {
-            FrequencyInner::CountDown(ref mut exec_count, _) => {
-                *exec_count -= 1u32;
-            }
-            FrequencyInner::Repeated(_) => {}
+            FrequencyInner::CronExpressionRepeated(_) => {}
+            FrequencyInner::SecondsRepeated(_) => {}
+            FrequencyInner::CronExpressionCountDown(ref mut exec_count, _) => *exec_count -= 1u64,
+            FrequencyInner::SecondsCountDown(count_down, _) => *count_down -= 1u64,
         };
     }
 
     fn is_down_over(&mut self) -> bool {
-        !matches!(self, FrequencyInner::CountDown(0, _))
+        matches!(
+            self,
+            FrequencyInner::CronExpressionCountDown(0, _) | FrequencyInner::SecondsCountDown(0, _)
+        )
     }
 }
 
@@ -328,7 +438,7 @@ impl FrequencyInner {
 /// Cycle plan task builder.
 pub struct TaskBuilder<'a> {
     /// Repeat type.
-    frequency: Frequency<'a>,
+    frequency: FrequencyUnify<'a>,
 
     /// Task_id should unique.
     task_id: u64,
@@ -430,35 +540,35 @@ pub struct Task {
     pub(crate) maximum_parallel_runnable_num: Option<u64>,
 }
 
-//bak type BoxFn
-// type BoxFn = Box<dyn Fn(i32) -> Pin<Box<dyn Future<Output = i32>>>>;
-
-enum RepeatType {
-    Num(u32),
-    Always,
-}
-
 impl<'a> TaskBuilder<'a> {
     /// Set task Frequency.
+    /// This api will be deprecated in the future, please use `set_frequency_once_*` | `set_frequency_count_down_*` | `set_frequency_repeated_*` etc.
+    #[deprecated]
     #[inline(always)]
     pub fn set_frequency(&mut self, frequency: Frequency<'a>) -> &mut Self {
-        self.frequency = frequency;
+        self.frequency = FrequencyUnify::FrequencyCronStr(frequency);
         self
     }
 
     /// Set task Frequency by customized CandyCronStr.
     /// In order to build a high-performance,
     /// highly reusable `TaskBuilder` that maintains the Copy feature .
+    ///
     /// when supporting building from CandyCronStr ,
     /// here actively leaks memory for create a str-slice (because str-slice support Copy, String does not)
+    ///
     /// We need to call `free` manually before `TaskBuilder` drop or before we leave the scope.
     ///
     /// Explain:
     /// Explicitly implementing both `Drop` and `Copy` trait on a type is currently
-    /// disallowed. This feature can make some sense in theory, but the current
+    /// disallowed.
+    ///
+    /// This feature can make some sense in theory, but the current
     /// implementation is incorrect and can lead to memory unsafety (see
     /// (issue #20126), so it has been disabled for now.
 
+    /// This api will be deprecated in the future, please use `set_frequency_*_by_candy` etc.
+    #[deprecated]
     #[inline(always)]
     pub fn set_frequency_by_candy<T: Into<CandyCronStr>>(
         &mut self,
@@ -474,12 +584,12 @@ impl<'a> TaskBuilder<'a> {
                 Frequency::Repeated(Box::leak(candy_cron_middle_str.into().0.into_boxed_str()))
             }
             CandyFrequency::CountDown(exec_count, candy_cron_middle_str) => Frequency::CountDown(
-                exec_count,
+                exec_count as u64,
                 Box::leak(candy_cron_middle_str.into().0.into_boxed_str()),
             ),
         };
 
-        self.frequency = frequency;
+        self.frequency = FrequencyUnify::FrequencyCronStr(frequency);
         self
     }
 
@@ -522,27 +632,7 @@ impl<'a> TaskBuilder<'a> {
     where
         F: Fn(TaskContext) -> Box<dyn DelayTaskHandler> + 'static + Send + Sync,
     {
-        let frequency_inner;
-
-        // The user inputs are pattern matched for different repetition types.
-        let (expression_str, repeat_type) = match self.frequency {
-            Frequency::Once(expression_str) => (expression_str, RepeatType::Num(1)),
-            Frequency::Repeated(expression_str) => (expression_str, RepeatType::Always),
-            Frequency::CountDown(exec_count, expression_str) => {
-                (expression_str, RepeatType::Num(exec_count))
-            }
-        };
-
-        let taskschedule = DelayTimerScheduleIteratorOwned::analyze_cron_expression(
-            self.schedule_iterator_time_zone,
-            expression_str,
-        )?;
-
-        // Building TaskFrequencyInner patterns based on repetition types.
-        frequency_inner = match repeat_type {
-            RepeatType::Always => FrequencyInner::Repeated(taskschedule),
-            RepeatType::Num(repeat_count) => FrequencyInner::CountDown(repeat_count, taskschedule),
-        };
+        let frequency_inner = (self.frequency, self.schedule_iterator_time_zone).try_into()?;
 
         let body = SafeStructBoxedFn(Box::new(body));
 
@@ -570,9 +660,10 @@ impl<'a> TaskBuilder<'a> {
     pub fn free(&mut self) {
         if self.build_by_candy_str {
             let s = match self.frequency {
-                Frequency::Once(s) => s,
-                Frequency::Repeated(s) => s,
-                Frequency::CountDown(_, s) => s,
+                FrequencyUnify::FrequencyCronStr(Frequency::Once(s)) => s,
+                FrequencyUnify::FrequencyCronStr(Frequency::Repeated(s)) => s,
+                FrequencyUnify::FrequencyCronStr(Frequency::CountDown(_, s)) => s,
+                _ => return,
             };
 
             unsafe {
@@ -582,6 +673,185 @@ impl<'a> TaskBuilder<'a> {
     }
 }
 
+impl<'a> TaskBuilder<'a> {
+    /// Task execution frequency: execute only once, set by cron expression.
+    #[inline(always)]
+    pub fn set_frequency_once_by_cron_str(&mut self, cron_str: &'a str) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencyCronStr(FrequencyCronStr::Once(cron_str));
+        self
+    }
+
+    /// Task execution frequency: countdown execution, set by cron expression.
+    #[inline(always)]
+    pub fn set_frequency_repeated_by_cron_str(&mut self, cron_str: &'a str) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencyCronStr(FrequencyCronStr::Repeated(cron_str));
+        self
+    }
+
+    /// Task execution frequency: execute repeatedly, set by cron expression.
+    #[inline(always)]
+    pub fn set_frequency_count_down_by_cron_str(
+        &mut self,
+        cron_str: &'a str,
+        count_down: u64,
+    ) -> &mut Self {
+        self.frequency =
+            FrequencyUnify::FrequencyCronStr(FrequencyCronStr::CountDown(count_down, cron_str));
+        self
+    }
+
+    /// Task execution frequency: execute only once, set by seconds num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_once_by_seconds(&mut self, seconds: u64) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::Once(seconds));
+        self
+    }
+
+    /// Task execution frequency: countdown execution, set by seconds num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_repeated_by_seconds(&mut self, seconds: u64) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::Repeated(seconds));
+        self
+    }
+
+    /// Task execution frequency: execute repeatedly, set by seconds num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_count_down_by_seconds(
+        &mut self,
+        seconds: u64,
+        count_down: u64,
+    ) -> &mut Self {
+        self.frequency =
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::CountDown(count_down, seconds));
+        self
+    }
+
+    /// Task execution frequency: execute only once, set by minutes num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    pub fn set_frequency_once_by_minutes(&mut self, minutes: u64) -> &mut Self {
+        self.frequency =
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::Once(ONE_MINUTE * minutes));
+        self
+    }
+
+    /// Task execution frequency: countdown execution, set by minutes num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_repeated_by_minutes(&mut self, minutes: u64) -> &mut Self {
+        self.frequency =
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::Repeated(ONE_MINUTE * minutes));
+        self
+    }
+
+    /// Task execution frequency: execute repeatedly, set by minutes num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_count_down_by_minutes(
+        &mut self,
+        minutes: u64,
+        count_down: u64,
+    ) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::CountDown(
+            count_down,
+            ONE_MINUTE * minutes,
+        ));
+        self
+    }
+
+    /// Task execution frequency: execute only once, set by hours num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    pub fn set_frequency_once_by_hours(&mut self, hours: u64) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::Once(ONE_HOUR * hours));
+        self
+    }
+
+    /// Task execution frequency: execute repeatedly, set by hours num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_repeated_by_hours(&mut self, hours: u64) -> &mut Self {
+        self.frequency =
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::Repeated(ONE_HOUR * hours));
+        self
+    }
+
+    /// Task execution frequency: countdown execution, set by hours num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_count_down_by_hours(&mut self, hours: u64, count_down: u64) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::CountDown(
+            count_down,
+            ONE_HOUR * hours,
+        ));
+        self
+    }
+
+    /// Task execution frequency: execute only once, set by days num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    pub fn set_frequency_once_by_days(&mut self, days: u64) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::Once(ONE_DAY * days));
+        self
+    }
+
+    /// Task execution frequency: execute repeatedly, set by days num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_repeated_by_days(&mut self, days: u64) -> &mut Self {
+        self.frequency =
+            FrequencyUnify::FrequencySeconds(FrequencySeconds::Repeated(ONE_DAY * days));
+        self
+    }
+
+    /// Task execution frequency: countdown execution, set by days num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    #[inline(always)]
+    pub fn set_frequency_count_down_by_days(&mut self, days: u64, count_down: u64) -> &mut Self {
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::CountDown(
+            count_down,
+            ONE_DAY * days,
+        ));
+        self
+    }
+
+    /// Task execution frequency: execute only once, set by timestamp-seconds num.
+    ///
+    /// Make sure time is greater than 1 seconds, otherwise undefined behavior will be triggered.
+
+    pub fn set_frequency_once_by_timestamp_seconds(&mut self, timestamp_seconds: u64) -> &mut Self {
+        let duration = timestamp_seconds
+            .checked_sub(get_timestamp())
+            .unwrap_or(ONE_SECOND);
+
+        self.frequency = FrequencyUnify::FrequencySeconds(FrequencySeconds::Once(duration));
+        self
+    }
+}
 impl Task {
     // get SafeBoxFn of Task to call.
     #[inline(always)]
@@ -598,16 +868,16 @@ impl Task {
         self.is_valid()
     }
 
-    //down_exec_count
+    // down_exec_count
     #[inline(always)]
     fn down_count(&mut self) {
         self.frequency.down_count();
     }
 
-    //set_valid_by_count_down
+    // set_valid_by_count_down
     #[inline(always)]
     fn set_valid_by_count_down(&mut self) {
-        self.valid = self.frequency.is_down_over();
+        self.valid = !self.frequency.is_down_over();
     }
 
     #[inline(always)]
@@ -672,18 +942,30 @@ impl Task {
     }
 }
 
+#[cfg(test)]
 mod tests {
-    #[allow(unused_imports)]
+    #![allow(deprecated)]
+
+    use super::{Task, TaskBuilder};
+    use crate::prelude::*;
+    use crate::utils::convenience::functions::create_default_delay_task_handler;
     use anyhow::Result as AnyResult;
+    use rand::prelude::*;
+    use std::iter::Iterator;
 
     #[test]
     fn test_task_valid() -> AnyResult<()> {
-        use super::{Frequency, Task, TaskBuilder};
-        use crate::utils::convenience::functions::create_default_delay_task_handler;
         let mut task_builder = TaskBuilder::default();
 
-        //The third run returns to an invalid state.
-        task_builder.set_frequency(Frequency::CountDown(3, "* * * * * * *"));
+        // The third run returns to an invalid state.
+        task_builder.set_frequency_count_down_by_seconds(1, 3);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        assert!(task.down_count_and_set_vaild());
+        assert!(task.down_count_and_set_vaild());
+        assert!(!task.down_count_and_set_vaild());
+
+        task_builder.set_frequency_count_down_by_cron_str("* * * * * * *", 3);
         let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
 
         assert!(task.down_count_and_set_vaild());
@@ -694,13 +976,150 @@ mod tests {
     }
 
     #[test]
-    fn test_is_can_running() -> AnyResult<()> {
-        use super::{Frequency, Task, TaskBuilder};
-        use crate::utils::convenience::functions::create_default_delay_task_handler;
+    fn test_get_next_exec_timestamp_seconds() -> AnyResult<()> {
+        let mut rng = rand::thread_rng();
+        let init_seconds: u64 = rng.gen_range(1..100_00_00);
         let mut task_builder = TaskBuilder::default();
 
-        //The third run returns to an invalid state.
-        task_builder.set_frequency(Frequency::CountDown(3, "* * * * * * *"));
+        task_builder.set_frequency_count_down_by_seconds(init_seconds, 3);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        (1..100)
+            .map(|i| {
+                debug_assert_eq!(
+                    task.get_next_exec_timestamp().unwrap(),
+                    get_timestamp() + (init_seconds * (i - 1))
+                );
+            })
+            .for_each(drop);
+
+        task_builder.set_frequency_count_down_by_cron_str("* * * * * * *", 100);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        (1..100)
+            .map(|_| {
+                assert!(task.down_count_and_set_vaild());
+            })
+            .for_each(drop);
+
+        assert!(!task.down_count_and_set_vaild());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_get_next_exec_timestamp_minutes() -> AnyResult<()> {
+        let mut rng = rand::thread_rng();
+        let init_minutes: u64 = rng.gen_range(1..100_00_00);
+        let mut task_builder = TaskBuilder::default();
+
+        task_builder.set_frequency_repeated_by_minutes(init_minutes);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        (1..100)
+            .map(|i| {
+                debug_assert_eq!(
+                    task.get_next_exec_timestamp().unwrap(),
+                    get_timestamp() + (init_minutes * (i - 1) * ONE_MINUTE)
+                );
+            })
+            .for_each(drop);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_get_next_exec_timestamp_hours() -> AnyResult<()> {
+        let mut rng = rand::thread_rng();
+        let init_hours: u64 = rng.gen_range(1..100_00_00);
+        let mut task_builder = TaskBuilder::default();
+
+        task_builder.set_frequency_repeated_by_hours(init_hours);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        (1..100)
+            .map(|i| {
+                debug_assert_eq!(
+                    task.get_next_exec_timestamp().unwrap(),
+                    get_timestamp() + (init_hours * (i - 1) * ONE_HOUR)
+                );
+            })
+            .for_each(drop);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_get_next_exec_timestamp_days() -> AnyResult<()> {
+        let mut rng = rand::thread_rng();
+        let init_days: u64 = rng.gen_range(1..100_00_00);
+        let mut task_builder = TaskBuilder::default();
+
+        task_builder.set_frequency_repeated_by_days(init_days);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        (1..100)
+            .map(|i| {
+                debug_assert_eq!(
+                    task.get_next_exec_timestamp().unwrap(),
+                    get_timestamp() + (init_days * (i - 1) * ONE_DAY)
+                );
+            })
+            .for_each(drop);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_count_down() -> AnyResult<()> {
+        let mut task_builder = TaskBuilder::default();
+
+        // The third run returns to an invalid state.
+        task_builder.set_frequency_count_down_by_seconds(1, 3);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        assert!(task.down_count_and_set_vaild());
+        assert!(task.down_count_and_set_vaild());
+        assert!(!task.down_count_and_set_vaild());
+
+        task_builder.set_frequency_count_down_by_cron_str("* * * * * * *", 3);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        assert!(task.down_count_and_set_vaild());
+        assert!(task.down_count_and_set_vaild());
+        assert!(!task.down_count_and_set_vaild());
+
+        task_builder.set_frequency_once_by_seconds(10);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+        assert!(!task.down_count_and_set_vaild());
+
+        task_builder.set_frequency_count_down_by_hours(10, 10);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+        (1i32..10i32)
+            .map(|_| assert!(task.down_count_and_set_vaild()))
+            .for_each(drop);
+        assert!(!task.down_count_and_set_vaild());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_is_can_running() -> AnyResult<()> {
+        let mut task_builder = TaskBuilder::default();
+
+        // The third run returns to an invalid state.
+        task_builder.set_frequency_count_down_by_cron_str("* * * * * * *", 3);
+        let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
+
+        assert!(task.is_can_running());
+
+        task.set_cylinder_line(1);
+        assert!(!task.is_can_running());
+
+        assert!(task.check_arrived());
+
+        // set_frequency_count_down_by_seconds.
+        task_builder.set_frequency_count_down_by_seconds(1, 1);
         let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;
 
         assert!(task.is_can_running());
@@ -713,15 +1132,12 @@ mod tests {
         Ok(())
     }
 
-    // struct CandyCron
-
     #[test]
     fn test_candy_cron() -> AnyResult<()> {
         use super::{CandyCron, CandyFrequency, Task, TaskBuilder};
-        use crate::utils::convenience::functions::create_default_delay_task_handler;
         let mut task_builder = TaskBuilder::default();
 
-        //The third run returns to an invalid state.
+        // The third run returns to an invalid state.
         task_builder.set_frequency_by_candy(CandyFrequency::CountDown(5, CandyCron::Minutely));
         let mut task: Task = task_builder.spawn(|_context| create_default_delay_task_handler())?;